+++ src/main/kotlin/io/hndrs/autoconfiguration/JwtArgumentResolverAutoconfiguration.kt
@@ -13,6 +13,7 @@ import io.hndrs.jwt.RequestTokenResolver
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication
 import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.boot.context.properties.ConstructorBinding
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -76,12 +77,13 @@ class JwtArgumentResolverAutoconfiguration(
 }
 
 @ConfigurationProperties(PROPERTY_PREFIX)
-class JwtConfigurationProperties {
-
+@ConstructorBinding
+data class JwtConfigurationProperties(
     /**
      * Keystore Path that needs to point to a valid jwks
      */
-    lateinit var keyStorePath: String
+    val keyStorePath: String
+) {
 
     companion object {
         const val PROPERTY_PREFIX = "hndrs.jwt"